Refactor Grant Application and Evaluation Models
- Changed GrantApplicationDAOImpl to manage Participants instead of GrantApplications.
- Updated EvaluationCreateDTO and EvaluationDTO to use Double for scores instead of Integer.
- Modified GrantApplication model to remove ownerId and adjusted related classes accordingly.
- Introduced Participant model to manage applications and added methods for adding/removing applications.
- Enhanced Evaluation model to reference GrantApplication directly.
- Updated service implementations to reflect changes in data models and ensure proper handling of evaluations.
- Improved test cases for registration, login, application creation, and evaluation processes.
- Added error handling for duplicate registrations and invalid login attempts.

+++ src/main/java/com/grantserver/dto/request/EvaluationCreateDTO.java
@@ -3,7 +3,7 @@ package com.grantserver.dto.request;
 public class EvaluationCreateDTO {
     public Long applicationId;
     public Long expertId;
-    public Integer score;
+    public Double score;
 
     public EvaluationCreateDTO() {}
 }
