+++ src/main/java/com/grantserver/service/impl/GrantFundServiceImpl.java
@@ -1,5 +1,9 @@
 package com.grantserver.service.impl;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import com.grantserver.common.config.ServiceRegistry;
 import com.grantserver.dao.EvaluationDAO;
 import com.grantserver.dao.GrantApplicationDAO;
@@ -11,10 +15,6 @@ import com.grantserver.model.GrantApplicationStatus;
 import com.grantserver.service.GrantFundService;
 import com.grantserver.service.algorithm.DefaultRankingStrategy;
 import com.grantserver.service.algorithm.RankingStrategy;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
 
 public class GrantFundServiceImpl implements GrantFundService {
 
@@ -32,32 +32,27 @@ public class GrantFundServiceImpl implements GrantFundService {
 
     @Override
     public GrantResultDTO calculate(GrantFundRequestDTO request) {
-        // 1. Получаем все АКТИВНЫЕ заявки
         List<GrantApplication> allApps = applicationDAO.findAll().stream()
                 .filter(app -> app.status == GrantApplicationStatus.ACTIVE)
                 .collect(Collectors.toList());
 
-        // 2. Считаем средний балл и оборачиваем в ScoredApplication
         List<RankingStrategy.ScoredApplication> scoredApps = new ArrayList<>();
         
         for (GrantApplication app : allApps) {
             List<Evaluation> evals = evaluationDAO.findByApplicationId(app.id);
             double avg = 0.0;
             if (!evals.isEmpty()) {
-                double sum = evals.stream().mapToInt(e -> e.score).sum();
+                double sum = evals.stream().mapToDouble(e -> e.score).sum();
                 avg = sum / evals.size();
             }
-            
-            // Фильтр по порогу (Threshold)
+
             if (avg >= request.threshold) {
                 scoredApps.add(new RankingStrategy.ScoredApplication(app, avg));
             }
         }
 
-        // 3. Сортировка по стратегии
         scoredApps.sort(rankingStrategy::compare);
 
-        // 4. Распределение фонда (Жадный алгоритм)
         List<GrantResultDTO.WinnerDTO> winners = new ArrayList<>();
         int currentFund = request.fund;
 
@@ -65,11 +60,9 @@ public class GrantFundServiceImpl implements GrantFundService {
             int requested = sa.application.requestedSum;
             
             if (currentFund >= requested) {
-                // Денег хватает - выдаем грант
                 currentFund -= requested;
                 winners.add(new GrantResultDTO.WinnerDTO(sa.application, sa.averageScore, requested));
             }
-            // Если денег не хватает, заявка пропускается
         }
 
         return new GrantResultDTO(winners, currentFund);
